CLEANUP: Dry up unwrap redundant org.logstash.ackedqueue.ext.JrubyAckedBatchExtLibrary
