+++ logstash-core/src/main/java/org/logstash/ackedqueue/ext/JrubyAckedQueueExtLibrary.java
@@ -154,7 +154,7 @@ public final class JrubyAckedQueueExtLibrary {
             }
 
             // TODO: return proper Batch object
-            return (b == null) ? context.nil : new JrubyAckedBatchExtLibrary.RubyAckedBatch(context.runtime, b);
+            return (b == null) ? context.nil : new RubyAckedBatch(context.runtime, b);
         }
 
         @JRubyMethod(name = "is_fully_acked?")
